requirements.txt
torch==2.8.0+cu134
torchvision==0.15.2+cu134
torchaudio==2.0.2
langchain
langchain-chroma
langchain-huggingface
gpt4all
unidecode
python-dotenv
openai
torch==2.8.0+cu134
torchvision==0.15.2+cu134
torchaudio==2.0.2
langchain==1.1.0
langchain-chroma==0.0.54
langchain-huggingface==0.0.10
gpt4all==1.1.2
unidecode==1.3.6
python-dotenv==1.0.0
openai==1.29.0
